pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "recall-cli"
version = "0.1.0"
description = "Local semantic memory search for your terminal"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Yajat Singh", email = "yajatns@gmail.com"}
]
keywords = ["memory", "semantic-search", "cli", "local", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "litellm>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
recall = "recall.cli:main"

[project.urls]
Homepage = "https://github.com/yajatns/recall"
Repository = "https://github.com/yajatns/recall"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
# Shell completion scripts are embedded strings with long lines
"src/recall/cli.py" = ["E501"]
